feat: Add support for Seeed T1000-E variant with LR1110 radio
- Introduced new board configuration for Seeed T1000-E in platformio.ini and JSON files.
- Added configuration header (config.h) for T1000-E, defining pin mappings and radio settings.
- Implemented LR1110 radio driver (radio_lr1110_nrf52.cpp) for nRF52840, ensuring compatibility with Rivr's API.
- Created BLE support for T1000-E, including rivr_ble_nrf52.cpp for Bluefruit integration.
- Updated heltec_t114 platformio.ini to include new BLE client environment.
- Enhanced variant files (variant.cpp and variant.h) to initialize pins and define digital/analog mappings.
- Added build flags and libraries necessary for the new T1000-E environment.

+++ rivr_layer/rivr_sinks.c
@@ -235,7 +235,8 @@ void rf_tx_sink_cb(const rivr_value_t *v, void *user_ctx)
             g_rivr_metrics.tx_queue_full++;
             ESP_LOGW(TAG, "rf_tx: queue full (fallback also failed) – dropped");
         } else {
-            RIVR_LOGI(TAG, "rf_tx: queued %u bytes (toa=%u us)", req.len, req.toa_us);
+            RIVR_LOGI(TAG, "rf_tx: queued %u bytes (toa=%lu us)",
+                     req.len, (unsigned long)req.toa_us);
             UPDATE_TXQ_PEAK();
             g_tx_frame_count++;            /* Track originated TX separately from relayed traffic so that a
              * heavy relay storm cannot exhaust the forward budget and silence
@@ -256,10 +257,11 @@ void rf_tx_sink_cb(const rivr_value_t *v, void *user_ctx)
     bool pushed = rb_try_push(&rf_tx_queue, &req);
     if (!pushed) {
         g_rivr_metrics.tx_queue_full++;
-        ESP_LOGW(TAG, "rf_tx: queue full – frame dropped (toa=%u us)", req.toa_us);
+        ESP_LOGW(TAG, "rf_tx: queue full – frame dropped (toa=%lu us)",
+                 (unsigned long)req.toa_us);
     } else {
-        RIVR_LOGI(TAG, "rf_tx: queued %u bytes (toa_approx=%u us, type=0x%02x)",
-                 req.len, req.toa_us, req.data[0]);
+        RIVR_LOGI(TAG, "rf_tx: queued %u bytes (toa_approx=%lu us, type=0x%02x)",
+                 req.len, (unsigned long)req.toa_us, req.data[0]);
         UPDATE_TXQ_PEAK();
         g_tx_frame_count++;
     }
